Chatbot/requirements.txt
openai==1.12.0
python-dotenv==1.0.0
streamlit==1.32.0
faiss-cpu==1.7.4
sentence-transformers==2.5.1
numpy==1.26.4
pandas==2.2.1 